% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_indicator.R
\name{add_fig_paths}
\alias{add_fig_paths}
\title{Simple function to add image file paths to indicator table template within targets pipeline}
\usage{
add_fig_paths(path, list_files)
}
\arguments{
\item{path}{The file path to the indicator table template (an .xlsx)}

\item{list_files}{Names of the images to include. This is a vector of text strings.}
}
\value{
A tibble
}
\description{
This function adds image file paths to the template data table indicator table template. Allows for tracking of the images and table data within the targets pipeline.
}
