% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recreational_indicators.R
\name{create_mrip_trips}
\alias{create_mrip_trips}
\title{Get MRIP trips file list
Create MRIP total recreational trips indicator}
\usage{
create_mrip_trips(files, remove_non_standard = TRUE)
}
\arguments{
\item{files}{A list of the full file names of annual directed trip data (.csv format). Must download for each year in MRIP query tool.}

\item{remove_non_standard}{Boolean, if TRUE will remove non-standard data ("Does Directed Trips Meet MRIP Standard" = NO)}
}
\value{
Saves R object \code{rec_trips}, returns directed recreational trips indicator
}
\description{
This function creates a total recreational trips indicator
For new data queries, use MRIP Query Tool (https://www.fisheries.noaa.gov/data-tools/recreational-fisheries-statistics-queries)
Query 'Directed Trip' under 'Effort Data'.
Choose year of interest, summarize by Annual, Calendar Year, Atlantic coast by state, species of interest, all modes and areas, Primary Target
Download csv as output
}
