% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recreational_indicators.R
\name{create_prop_sp_trips}
\alias{create_prop_sp_trips}
\title{Create MRIP species recreational effort indicator}
\usage{
create_prop_sp_trips(
  total_trips,
  species_trips,
  states = c("MAINE", "CONNECTICUT", "MASSACHUSETTS", "NEW HAMPSHIRE", "NEW JERSEY",
    "NEW YORK", "RHODE ISLAND", "MARYLAND", "DELAWARE", "NORTH CAROLINA"),
  groupby_state = FALSE,
  return = TRUE
)
}
\arguments{
\item{total_trips}{The MRIP trip data (R object \code{mrip_effort}) of total annual angler trips from downloaded csv file}

\item{species_trips}{The directed trips data for species of interest (R object \code{rec_trips})}

\item{states}{States in which to filter data, from MRIP query 'ATLANTIC COAST BY STATE'}

\item{groupby_state}{Boolean, if TRUE will group by state, if FALSE will sum across all states}

\item{return}{Boolean, if TRUE will return the R data object \code{prop_sp_trips}}
}
\value{
Saves the R data object \code{prop_sp_trips}
}
\description{
This function creates an indicator with the proportion trips targeting a species of interest
For new data queries, use MRIP Query Tool (https://www.fisheries.noaa.gov/data-tools/recreational-fisheries-statistics-queries)
Query 'Time Series' under 'Effort Data'.
Choose years of interest, summarize by Annual, Calendar Year, Atlantic Coast by State, all modes and areas
Download csv as output
}
