% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bottomtrawl_indicators.R
\name{create_stratified_mean}
\alias{create_stratified_mean}
\title{Calculate Stratified Mean Biomass}
\usage{
create_stratified_mean(...)
}
\arguments{
\item{...}{passed to \code{survdat::calc_stratified_mean()}. Parameters here: https://noaa-edab.github.io/survdat/reference/calc_stratified_mean.html}
}
\value{
a data table 'strat_mean'
}
\description{
This function calculates stratified mean biomass
}
