% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bottomtrawl_indicators.R
\name{create_swept_area}
\alias{create_swept_area}
\title{Calculate Swept Area Biomass}
\usage{
create_swept_area(...)
}
\arguments{
\item{...}{passed to 'survdat::calc_swept_area()`}
}
\value{
a data table 'swept_area'
}
\description{
This function calculates swept area biomass
}
