% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{Create an ESP template}
\usage{
create_template(path = getwd(), over = FALSE)
}
\arguments{
\item{path}{Where to create the template. Defaults to the present working directory.}

\item{over}{whether to overwrite existing files. Defaults to FALSE.}
}
\description{
This function creates an ESP two-pager template at the specified path.
}
