% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recreational_indicators.R
\name{create_total_rec_catch}
\alias{create_total_rec_catch}
\title{Create MRIP total recreational catch indicator}
\usage{
create_total_rec_catch(
  data,
  states = c("MAINE", "CONNECTICUT", "MASSACHUSETTS", "NEW HAMPSHIRE", "NEW JERSEY",
    "NEW YORK", "RHODE ISLAND", "MARYLAND", "DELAWARE", "NORTH CAROLINA"),
  groupby_state = FALSE,
  return = TRUE
)
}
\arguments{
\item{data}{The mrip data (R object \code{mrip_catch}), already subset to species of interest only from downloaded csv file}

\item{states}{States in which to filter data, from MRIP query 'ATLANTIC COAST BY STATE'}

\item{groupby_state}{Whether to group results by state}
}
\value{
Saves the R data object \code{total_rec_catch}
}
\description{
This function creates the total recreational catch indicator
For new data queries, use MRIP Query Tool (https://www.fisheries.noaa.gov/data-tools/recreational-fisheries-statistics-queries)
Query 'Time Series' under 'Catch Data'.
Choose years of interest, summarize by Annual, Calendar Year, Atlantic coast by state, species of interest, all modes and areas, Total Catch
Download csv as output
}
