% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrip_functions.R
\name{create_total_rec_landings}
\alias{create_total_rec_landings}
\title{Create MRIP recreational landings indicator}
\usage{
create_total_rec_landings(
  data,
  states = c("MAINE", "CONNECTICUT", "MASSACHUSETTS", "NEW HAMPSHIRE", "NEW JERSEY",
    "NEW YORK", "RHODE ISLAND", "MARYLAND", "DELAWARE", "NORTH CAROLINA"),
  groupby_state = FALSE,
  return = TRUE
)
}
\arguments{
\item{data}{The MRIP harvest data (R object \code{mrip_landing})}

\item{states}{States in which to filter data, from MRIP query 'ATLANTIC COAST BY STATE'}
}
\value{
Saves the R data object \code{total_rec_landings}
}
\description{
This function creates an indicator for total recreational landings
For new data queries, use MRIP Query Tool (https://www.fisheries.noaa.gov/data-tools/recreational-fisheries-statistics-queries)
Query 'Time Series' under 'Catch Data'.
Choose years of interest, summarize by Annual, Calendar Year, Atlantic Coast by State, species of interest, all modes and areas, Harvest (A + B1), choose # of fish/weight (lbs), mean length and weight
Download csv as output
}
