% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bottomtrawl_indicators.R
\name{diet}
\alias{diet}
\title{Create a table of diet data}
\usage{
diet(data)
}
\arguments{
\item{data}{A data frame from \code{allfh}.}
}
\value{
A \code{DT::datatable} or a \code{knitr::kable}
}
\description{
Create a table of diet data
}
