% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_netcdf.R
\name{esp_csv_to_nc}
\alias{esp_csv_to_nc}
\title{Convert longform tidy ESP indicator data to netcdf}
\usage{
esp_csv_to_nc(data, fname)
}
\arguments{
\item{data}{A data frame containing ESP indicator data in long format. Must contain columns: INDICATOR_NAME, UNITS, DATA_VALUE, YEAR, and SUBMISSION_YEAR}

\item{fname}{A character string specifying the file path where the .nc file will be saved}
}
\value{
.nc file saved at the specified filepath
}
\description{
This function convertd longform tidy ESP indicator data to netcdf
}
