% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_indicator.R
\name{format_from_template}
\alias{format_from_template}
\title{Wrapper function to format data for upload to AKFIN web service}
\usage{
format_from_template(key, ind_name, ...)
}
\arguments{
\item{key}{The metadata (a tibble)}

\item{ind_name}{Name of the indicator. This is a primary key and must match what has been used in previous years. This is a text string.}

\item{...}{Passed to \code{NEesp2::format_indicator()}}
}
\value{
A .txt file saved in \code{out_dir} (\code{out_dir} must be pased with \code{...})
}
\description{
This function is a wrapper that allows a metadata data table to be used to format indicator data for upload to AKFIN web service
}
