% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_indicator.R
\name{format_indicator}
\alias{format_indicator}
\title{Format data for upload to AKFIN web service}
\usage{
format_indicator(
  out_dir,
  submission_year,
  indicator_name,
  description,
  status,
  factors,
  implications,
  references,
  years,
  region,
  indicator_value,
  ...
)
}
\arguments{
\item{out_dir}{dictory where the .txt file will be saved}

\item{submission_year}{Current year of contribution submission. This is a integer value.}

\item{indicator_name}{Name of the indicator. This is a primary key and must match what has been used in previous years. This is a text string.}

\item{description}{Brief description of the indicator. Please make sure this description includes information on the spatial distribution of the indicator and how the data for the indicator are collected. The spatial resolution can be a cell size dimension (e.g., 5 km resolution for gridded data) or area of sampling for a survey. The data collection method can be brief (e.g., survey name and gear type, satellite sensor and version, stock assessment model output, fishery observer data, community reports, etc.) and can include a reference to methods detailed elswhere. This is a text string.}

\item{status}{Information on the current status of the indicator in the context of historical trends. This is a text string.}

\item{factors}{Information on the potential causes for observed trends and current status (focus on most recent year). This is a text string.}

\item{implications}{Information that briefly answers these questions: What are the implications or impacts of the observed trends on the ecosystem or ecosystem components? What do the trends mean? Why are they important? How can this information be used to inform management decisions? This is a text string.}

\item{references}{Include any full references that are associated with the indicator. This may include data references such as from an ERDDAP webpage or literature cited (plus the DOI where possible). This is a text string.}

\item{years}{Years for the indicator contribution. This is a numeric vector.}

\item{region}{List of spatial or temporal scales applicable to indicator. This is a character value.}

\item{indicator_value}{Data values for the indicator contribution (must match the YEAR list length). This is a numeric vector.}

\item{...}{not used}
}
\value{
A .txt file saved in \code{out_dir}
}
\description{
This function formats indicator data for upload to AKFIN web service
}
