% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_report_card.R
\name{format_tbl_data}
\alias{format_tbl_data}
\title{Format data from report card spreadsheet}
\usage{
format_tbl_data(file, term_year, dir)
}
\arguments{
\item{file}{the file path}

\item{term_year}{the terminal year to use in the table header}

\item{dir}{the directory where the time series images are saved}
}
\value{
A tibble
}
\description{
This function formats the data to prep for flextable creation.
}
