% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrip_data.R
\name{get_mrip_trips}
\alias{get_mrip_trips}
\title{Scrape MRIP trip data from MRIP Query tool}
\usage{
get_mrip_trips(species, region, year)
}
\arguments{
\item{species}{the common name of the species as it appears in the MRIP data. capitalization does not matter.}

\item{region}{the name of the region. Can be a state name, "North Atlantic", "Mid-Atlantic", etc. Capitalization does not matter.}

\item{year}{the year of data to query. Must be a single value. The earliest year possible is 1981.}
}
\value{
Returns a list of the scraped data and metadata.
}
\description{
This function scrapes MRIP trip data from the MRIP Query tool
}
