% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recreational_indicators.R
\name{groupby_state}
\alias{groupby_state}
\title{Utility function to group by}
\usage{
groupby_state(data, groupby)
}
\arguments{
\item{data}{the data (a tibble)}

\item{groupby}{Boolean, if TRUE will group by STATE and YEAR, if false will group by YEAR only}
}
\description{
This function groups by year or by year and state
}
