% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hms_key}
\alias{hms_key}
\title{Highly Migratory Species (HMS) Names and Species Codes}
\format{
A data frame with 42 rows and 4 variables:
\describe{
\item{\code{SP_CODE}}{double Species code numbers to create the rec_hms indicator in ecodata}
\item{\code{COMMON_NAME}}{character Common name of the species}
\item{\code{SP_CATEGORY}}{character HMS Category (Large Coastal, Small Coastal, Prohibited, Pelagic, Scombridae, Billfishes)}
\item{\code{COMMON_POP}}{character Other identifier for some shark species}
}
}
\usage{
hms_key
}
\description{
A list of HMS species in the North Atlantic with corresponding species code numbers and categories.
}
\keyword{datasets}
