% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_ts.R
\name{plt_indicator}
\alias{plt_indicator}
\title{Plot indicator time series}
\usage{
plt_indicator(data)
}
\arguments{
\item{data}{A data frame with columns of YEAR, INDICATOR_NAME, and DATA_VALUE}
}
\value{
A ggplot
}
\description{
This function plots an indicator time series with a year, indicator name, and data value with mean and standard deviation lines.
}
