% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrip_functions.R
\name{read_rec_catch}
\alias{read_rec_catch}
\title{Read in MRIP recreational catch data}
\usage{
read_rec_catch(species, dir, type = "all")
}
\arguments{
\item{species}{the species common name}

\item{dir}{the directory where MRIP catch files are saved}

\item{type}{the type of data to read in, options are "all", "private", "charter", or "for_hire"}
}
\value{
a tibble
}
\description{
This function read in MRIP catch data
}
