% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrip_data.R
\name{save_catch}
\alias{save_catch}
\title{Scrape and save MRIP catch data from MRIP Query tool}
\usage{
save_catch(this_species, out_folder)
}
\arguments{
\item{this_species}{the common name of the species as it appears in the MRIP data. capitalization does not matter.}

\item{out_folder}{where to save the data}
}
\value{
Returns a list of the scraped data and metadata.
}
\description{
This function scrapes MRIP catch data from the MRIP Query tool and saves it as an Rds.
Used as a helper function to automate data pulls.
}
