% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species.codes}
\alias{species.codes}
\title{Species codes for bottom trawl survey}
\format{
A data frame with 50 rows and 2 variables:
\describe{
\item{\code{SVSPP}}{integer A standard code which represents a species caught in a trawl or dredge. Refer to the SVDBS.SVSPECIES_LIST}
\item{\code{Species}}{character Common name of the species}
}
}
\usage{
species.codes
}
\description{
A list of species codes (SVSPP) and their corresponding common names as found in a bottom trawl survey data pull.
}
\keyword{datasets}
