% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bottomtrawl_indicators.R
\name{species_condition}
\alias{species_condition}
\title{Calculate Species Condition (weight/length^3)}
\usage{
species_condition(data)
}
\arguments{
\item{data}{A data frame from \code{allfh}.}
}
\value{
a data frame
}
\description{
This function calculates condition (Fulton's K)
}
