% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_indicator_old.R
\name{species_condition_old}
\alias{species_condition_old}
\title{Create species condition indicator}
\usage{
species_condition_old(
  data,
  LWparams = NEesp2::LWparams,
  species.codes = NEesp2::species.codes,
  by_EPU = TRUE,
  by_sex = FALSE,
  length_break = NULL,
  more_than_20_years = TRUE,
  record_outliers = FALSE,
  output = "soe"
)
}
\arguments{
\item{data}{data table. NEFSC survey data generated by 'survdat::get_survdat_data(channel, getBio = TRUE, getLengths = TRUE)'}

\item{LWparams}{data table. L-W parameters. Default is from Wigley et al. 2003 from NEesp2::LWparams. Alternative parameters could be used, as long as they were formatted the same as the Wigley parameters.}

\item{species.codes}{data table. Species codes (SVSPP) and common names from NEesp2::species.codes}

\item{by_EPU}{logical. If TRUE, calculates condition by EPUs specified in the input data, if FALSE, calculates condition for all data combined.}

\item{by_sex}{logical. If TRUE, calculates condition by sex. If FALSE, calculates condition across sexes.}

\item{length_break}{numeric vector. If not NULL, will calculate condition by length breaks specified in the vector. User must specify minimum and maximum lengths in this parameter, e.g., c(0, 20, 70). If NULL, will not calculate by length groupings.}

\item{more_than_20_years}{logical. If TRUE, only returns species with 20 or more years of data.}

\item{record_outliers}{logical. If TRUE, returns a list with two data frames: the first is the condition data frame, the second is a data frame of outliers that were removed from the analysis. If FALSE, only returns the condition data frame.}

\item{output}{character. If "soe", returns a data frame of species condition for the State of the Ecosystem report. If "esp", returns a data frame for ESPs. If "full", returns a data frame of all calculated values. \emph{Setting by_sex = TRUE or length_break to any value will always return a full dataframe}}
}
\value{
Returns a data frame of species condition
}
\description{
This function calculates species condition from a survdat data pull.
Methods derived from Laurel Smith (https://github.com/Laurels1/Condition/blob/master/R/RelConditionEPU.R)
}
