% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{DATASET_TITLE}
\format{
A data frame with 2047 rows and 21 variables:
\describe{
\item{\code{SCINAME}}{character COLUMN_DESCRIPTION}
\item{\code{COMNAME}}{character COLUMN_DESCRIPTION}
\item{\code{SVSPP}}{integer COLUMN_DESCRIPTION}
\item{\code{SVABBR}}{character COLUMN_DESCRIPTION}
\item{\code{PYSPP}}{double COLUMN_DESCRIPTION}
\item{\code{PYABBR}}{character COLUMN_DESCRIPTION}
\item{\code{NODCCODE}}{character COLUMN_DESCRIPTION}
\item{\code{NODCTYPE}}{logical COLUMN_DESCRIPTION}
\item{\code{NODCLEVL}}{character COLUMN_DESCRIPTION}
\item{\code{NODCCTRL}}{character COLUMN_DESCRIPTION}
\item{\code{NODCAUTH}}{character COLUMN_DESCRIPTION}
\item{\code{AUTHOR}}{character COLUMN_DESCRIPTION}
\item{\code{PICTURE}}{character COLUMN_DESCRIPTION}
\item{\code{COMMNT}}{character COLUMN_DESCRIPTION}
\item{\code{O_SVSCNM}}{character COLUMN_DESCRIPTION}
\item{\code{O_SVCONM}}{character COLUMN_DESCRIPTION}
\item{\code{O_SVSPP}}{integer COLUMN_DESCRIPTION}
\item{\code{DOC}}{character COLUMN_DESCRIPTION}
\item{\code{UOC}}{character COLUMN_DESCRIPTION}
\item{\code{DOE}}{character COLUMN_DESCRIPTION}
\item{\code{UOE}}{character COLUMN_DESCRIPTION}
}
}
\usage{
species_data
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
