% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata_epu_key}
\alias{strata_epu_key}
\title{Strata EPU Key}
\format{
A data frame with 1553 rows and 2 variables:
\describe{
\item{\code{STRATUM}}{integer Bottom trawl survey strata number}
\item{\code{EPU}}{character Ecological Production Unit (Mid-Atlantic Bight - MAB, Georges Bank - GB, Gulf of Maine - GOM, Scotian Shelf -SS)}
}
}
\usage{
strata_epu_key
}
\description{
A list of bottom trawl survey strata and their corresponding EPUs on the Northeast U.S. Continental Shelf.
}
\keyword{datasets}
