% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survdat_subset}
\alias{survdat_subset}
\title{Survdat data pull}
\format{
A data frame with 16527 rows and 28 variables:
\describe{
\item{\code{CRUISE6}}{integer Code uniquely identifying cruise. The first four digits indicate the year and the last two digit uniquely identify the cruise within the year.}
\item{\code{STATION}}{integer Unique sequential order in which stations have been completed.}
\item{\code{STRATUM}}{integer A predefined area where a net dredge, or other piece of gear was deployed. Code consists of 2 parts: Stratum group code number (2 bytes) and stratum number (3 bytes).}
\item{\code{SVSPP}}{integer A standard code which represents a species caught in a trawl or dredge. Refer to the SVDBS.SVSPECIES_LIST}
\item{\code{CATCHSEX}}{integer Code used to identify species that are sexed at the catch level.}
\item{\code{LENGTH}}{integer Measured length of species in centimeters (cm).}
\item{\code{TOW}}{integer Sequential number representing order in which station was selected within a stratum.}
\item{\code{SVVESSEL}}{character Standard two character code for a survey vessel.}
\item{\code{YEAR}}{integer Year in which cruise was conducted.}
\item{\code{SEASON}}{character Season of the year in which cruise was conducted.}
\item{\code{LAT}}{double Beginning latitude of tow in decimal degrees.}
\item{\code{LON}}{double Beginning longitude of tow in decimal degrees.}
\item{\code{EST_TOWDATE}}{character Date and time represented by Eastern Standard Time (EST) for the start of a tow or deployment.}
\item{\code{DEPTH}}{integer A four digit number recording the average depth, to the nearest meter, during a survey gear deployment.}
\item{\code{SURFTEMP}}{double Surface temperature of water (degrees Celsius).}
\item{\code{SURFSALIN}}{double Salinity at water surface in practical salinity units (PSU).}
\item{\code{BOTTEMP}}{double Bottom temperature (degrees Celsius).}
\item{\code{BOTSALIN}}{double Bottom salinity in Practical Salinity Units (PSU).}
\item{\code{ABUNDANCE}}{integer Expanded number of individuals of a species caught at a given station.}
\item{\code{BIOMASS}}{double Expanded catch weight of a species caught at a given station.}
\item{\code{NUMLEN}}{integer Expanded number of specimens at a given length.}
\item{\code{INDID}}{integer A unique identifier for each fish sampled.}
\item{\code{INDWT}}{double Individual weight (KG) of species being sampled.}
\item{\code{SEX}}{integer Code indicating sex of fish or invertebrate species.}
\item{\code{MATURITY}}{character Stage of maturation of the fish being sampled.}
\item{\code{AGE}}{integer Age of specimen in years.}
\item{\code{STOM_VOLUME}}{double Volume of the stomach contents of the fish sampled, measured to the nearest tenth of a cubic centimeter (cc).}
\item{\code{STOM_WGT}}{double Stomach weight of an individual fish in grams.}
}
}
\usage{
survdat_subset
}
\description{
A data pull of bottom trawl survey data from 1983-2024 using the survdat package.
}
\keyword{datasets}
