% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_indicator.R
\name{create_spatial_indicator}
\alias{create_spatial_indicator}
\title{Create Spatial Indicator}
\usage{
create_spatial_indicator(indicator_name, units, ...)
}
\arguments{
\item{indicator_name}{name of the indicator (character)}

\item{units}{units associated with the indicator (character)}

\item{...}{passed to \code{EDABUtilities::make_2d_summary_ts}}
}
\value{
Returns a data frame summarized by timestep for each area.names
}
\description{
This function generates a spatially-aggregated indicator from a netCDF file.
The function passes data through EDABUtilities::make_2d_summary_ts, which provides summary statistics of 2d gridded data as time series by area.
Converts .nc files to data frame.
}
