% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bottomtrawl_indicators.R
\name{create_swept_area}
\alias{create_swept_area}
\title{Calculate Swept Area Biomass}
\usage{
create_swept_area(...)
}
\arguments{
\item{...}{passed to \code{survdat::calc_swept_area()}. Parameters here: https://noaa-edab.github.io/survdat/reference/calc_swept_area.html}
}
\value{
a data table 'swept_area'
}
\description{
This function calculates swept area biomass
}
