% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recreational_indicators.R
\name{create_total_rec_catch}
\alias{create_total_rec_catch}
\title{Create MRIP total recreational catch indicator}
\usage{
create_total_rec_catch(data, var_name = "catch", remove_non_standard = TRUE)
}
\arguments{
\item{data}{The mrip data}

\item{var_name}{The variable name to use in the indicator name. Default is "catch".}

\item{remove_non_standard}{Boolean, if TRUE will remove non-standard data ("Does Total Catch (A+B1+B2) Meet MRIP Standard" = NO)}

\item{species}{The species common name}
}
\value{
a tibble
}
\description{
This function creates the total recreational catch indicator
Input data is MRIP catch (A, B1, B2 catch combined)
}
