% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrip_data.R
\name{get_mrip_catch}
\alias{get_mrip_catch}
\title{Scrape MRIP catch data from MRIP Query tool}
\usage{
get_mrip_catch(species)
}
\arguments{
\item{species}{the common name of the species as it appears in the MRIP data. capitalization does not matter.}
}
\value{
Returns a list of the scraped data and metadata.
}
\description{
This function scrapes MRIP catch data from the MRIP Query tool
}
