% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrip_functions.R
\name{get_trip_files}
\alias{get_trip_files}
\title{This function returns a list of MRIP trip files}
\usage{
get_trip_files(dir, species)
}
\arguments{
\item{dir}{A directory that has subfolders with species-level data}

\item{species}{the species of interest}
}
\value{
A vector of files
}
\description{
This function returns a list of MRIP trip files
}
