% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition.R
\name{plot_condition}
\alias{plot_condition}
\title{Plot Condition}
\usage{
plot_condition(data, var, return = TRUE)
}
\arguments{
\item{data}{A data frame of species condition from 'species_condition' function}

\item{var}{species of interest to plot}
}
\value{
A ggplot
}
\description{
Plot Condition
}
