% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_report_card.R
\name{rpt_card_table}
\alias{rpt_card_table}
\title{Create a flextable with inset indicator time series plots}
\usage{
rpt_card_table(data, widths = c(0.9, 0.75, 3, 3))
}
\arguments{
\item{data}{A data frame or tibble}

\item{widths}{A numeric vector of column widths (in inches). Default is c(0.9, 0.75, 3, 3)}
}
\value{
A flextable
}
\description{
This function creates a flextable for ESP report cards.
}
