% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrip_data.R
\name{save_catch}
\alias{save_catch}
\title{Scrape and save MRIP catch data from MRIP Query tool}
\usage{
save_catch(
  this_species,
  this_region,
  out_folder,
  catch_type = "all",
  this_data_type,
  wait = TRUE,
  return_fname = TRUE
)
}
\arguments{
\item{this_species}{the common name of the species as it appears in the MRIP data. capitalization does not matter.}

\item{out_folder}{where to save the data}

\item{catch_type}{the type of catch to query. Can be "all" for all catch types (A, B1, B2), or "landings" for just the landings (A and B1). Default is "all".}

\item{wait}{whether to pause after saving the data. Default is TRUE.}

\item{return_fmane}{whether to return the file name of the saved data. Default is TRUE.}
}
\value{
Saves list of the scraped data and metadata. Returns the file name.
}
\description{
This function scrapes MRIP catch data from the MRIP Query tool and saves it as an Rds.
Used as a helper function to automate data pulls.
}
