% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_indicator.R
\name{species_condition}
\alias{species_condition}
\title{Create species condition indicator}
\usage{
species_condition(
  data,
  LWparams = NEesp2::LWparams,
  species.codes = NEesp2::species.codes,
  by_EPU = TRUE
)
}
\arguments{
\item{data}{data table. NEFSC survey data generated by 'survdat::get_survdat_data(channel, getBio = TRUE, getLengths = TRUE)'}

\item{LWparams}{data table. L-W parameters. Default is from Wigley et al. 2003 from NEesp2::LWparams. Alternative parameters could be used, as long as they were formatted the same as the Wigley parameters.}

\item{species.codes}{data table. Species codes (SVSPP) and common names from NEesp2::species.codes}

\item{by_EPU}{logical. If TRUE, calculates condition by EPUs specified in the input data, if FALSE, calculates condition for all data combined.}
}
\value{
Returns a data frame of species condition
}
\description{
This function calculates species condition from a survdat data pull.
Methods derived from Laurel Smith (https://github.com/Laurels1/Condition/blob/master/R/RelConditionEPU.R)
}
