% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_indicator.R
\name{species_condition}
\alias{species_condition}
\title{Create species condition indicator}
\usage{
species_condition(data, LWparams, species.codes)
}
\arguments{
\item{data}{data table. NEFSC survey data generated by 'survdat::get_survdat_data(channel, getBio = T, getLengths = T)'}

\item{LWparams}{L-W parameters from Wigley et al. 2003 from 'data-raw/tech_memo_parameters_table_format.csv'}

\item{species.codes}{data table. Species codes and names from 'data-raw/bottomtrawl_species_codes_names.csv'}
}
\value{
Returns R object \code{condition}, a data frame of species condition by EPU with scale.
}
\description{
This function calculates species condition from a survdat data pull.
Methods derived from Laurel Smith (https://github.com/Laurels1/Condition/blob/master/R/RelConditionEPU.R)
}
