% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{Species lookup table from {survdat}}
\format{
A data frame with 2047 rows and 21 variables:
\describe{
\item{\code{SCINAME}}{character Scientific name of specimen.}
\item{\code{COMNAME}}{character Accepted common name of a fish or invertebrate species.}
\item{\code{SVSPP}}{integer A standard code which represents a species caught in a trawl or dredge.}
\item{\code{SVABBR}}{character Abbreviation of common name of specimen.}
\item{\code{PYSPP}}{double Prey species name.}
\item{\code{PYABBR}}{character Prey abbreviation.}
\item{\code{NODCCODE}}{character National Oceanographic Data Center (NODC) taxononic code for species.}
\item{\code{NODCTYPE}}{logical National Oceanographic Data Center (NODC) Type name code: S-scientific name, C-Common name}
\item{\code{NODCLEVL}}{character National Oceanographic Data Center (NODC) taxanomic level: P-Phylum, C-Class, O-Order, F-Family, G-Genus, S-Species}
\item{\code{NODCCTRL}}{character National Oceanographic Data Center (NODC).}
\item{\code{NODCAUTH}}{character NODC code author (person who described species).}
\item{\code{AUTHOR}}{character Author of document.}
\item{\code{PICTURE}}{character Citation for picture of species.}
\item{\code{COMMNT}}{character Comment related to species in SVDBS.SVSPECIES_LIST.}
\item{\code{O_SVSCNM}}{character Former scientific name used by Survey branch.}
\item{\code{O_SVCONM}}{character Former common name used by Survey branch.}
\item{\code{O_SVSPP}}{integer Former species code.}
\item{\code{DOC}}{character Date when the current record was updated or changed. Date format (MM/DD/YY HH:MI:SS)}
\item{\code{UOC}}{character The Oracle username of the individual who changed the current record.}
\item{\code{DOE}}{character Date when record was inserted into Oracle. Date format (MM/DD/YY HH:MI:SS).}
\item{\code{UOE}}{character The Oracle username of the individual who entered the new record.}
}
}
\usage{
species_data
}
\description{
A species lookup table generated from survdat::get_species()
}
\keyword{datasets}
