% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bottomtrawl_indicators.R
\name{species_range}
\alias{species_range}
\title{Calculate Species Range}
\usage{
species_range(data, species)
}
\arguments{
\item{data}{data table. NEFSC survey data generated by 'survdat::get_survdat_data(channel)'}

\item{species}{data table. NEFSC species data generated by 'survdat::get_species(channel)'}
}
\value{
a data frame
}
\description{
Calculate Species Range
}
